GetLanguages and GetStatuses migrated to info.go
